[refactor] - 코드 리뷰 적용중
