+++ src/components/room/lobby/OtherRoomsTitle.tsx
@@ -5,7 +5,7 @@ function OtherRoomsTitle({ children }: { children: React.ReactNode }) {
   return (
     <article className="z-10 pt-[454px]">
       <div className="flex flex-col justify-start max-w-[1000px]">
-        <div className="text-[40px]	font-semibold">모집 중</div>
+        <p className="text-[40px]	font-semibold">모집 중</p>
         <div className="flex flex-row gap-x-[16px] mt-[24px] w-1/4">
           <RegionSelection text={'selectRegion'} />
           <MemberNumberSelection text={'selectMember'} />
